fix: a registry from a Perch far enough ahead was reported as a corrupt file
`claimed_version` read `version` into a `u32`, so a number past that ceiling
deserialized as nothing at all. `says_no_version` is false for it — it is not
null — so the guard fell through, the migration declined, and the parse failed
with serde's words about an integer that will not fit, wrapped as `Malformed`
with instructions to hand-edit the file. The one thing a version exists to catch
was the one thing it reported as corruption.

Read as a `u64` now and compared against the ceiling widened to match, so the
answer is "written by a newer Perch — Upgrade Perch", naming the version it
actually read. A `version` that is no whole number is still serde's to describe,
which is the existing call and the right one: it names the value.

+++ src/export.rs
@@ -441,7 +441,7 @@ fn refuse_a_newer_perch(plain: &[u8]) -> Result<()> {
         return Err(crate::error::written_by_a_newer_perch(
             "This Export",
             "export",
-            versioned.version.unwrap_or_default(),
+            u64::from(versioned.version.unwrap_or_default()),
             CURRENT_VERSION,
         ));
     }
@@ -454,7 +454,7 @@ fn refuse_a_newer_perch(plain: &[u8]) -> Result<()> {
         return Err(crate::error::written_by_a_newer_perch(
             "The registry inside this Export",
             "registry",
-            inside.unwrap_or_default(),
+            u64::from(inside.unwrap_or_default()),
             crate::registry::CURRENT_VERSION,
         ));
     }
